fix int overflow bug

+++ app/src/main/java/com/elzren/mywatchlist/mediaDetail/domain/model/MovieDetail.kt
@@ -8,7 +8,7 @@ data class MovieDetail(
     @SerializedName("backdrop_path")
     val backdropPath: String?,
     @SerializedName("budget")
-    val budget: Int,
+    val budget: Long,
     @SerializedName("genres")
     val genres: List<Genre>,
     @SerializedName("homepage")
@@ -30,7 +30,7 @@ data class MovieDetail(
     @SerializedName("release_date")
     val releaseDate: String,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("runtime")
     val runtime: Int,
     @SerializedName("status")
